requirements.txt
Flask==2.0.3
numpy==1.21.5
Pillow==10.1.0
tensorflow==2.6.0
